\name{elementR}
\docType{package}
\title{
A Shiny Application for Reducing Elemental LA-ICPMS Data from Solid Structures 
}
\description{
Aims to facilitate the reduction of elemental microchemistry data from solid-phase LA-ICPMS analysis (laser ablation inductive coupled plasma mass spectrometry). The elementR package provides a reactive and user friendly interface for conducting all steps needed for an optimal data reduction while leaving maximum control for user.

}
\section{Author}{
Charlotte Sirot, Francois Guilhaumon, Franck Ferraton, Audrey Darnaude, Jacques Panfili, Amber Child
}

\section{Maintainer}{
Charlotte Sirot <charlott.sirot@gmail.com>
}

\references{
Calculations of elementR procedures are based on a consensus of the literature:
\itemize{
\item {Elsdon & Gillanders. Interactive effects of temperature and salinity on otolith chemistry: challenges for determining environmental histories of fish. Can. J. Fish. Aquat. Sci. Vol. 59, 2002.}
\item{Fowler et al. Experimental assessment of the effect of temperature and salinity on elemental composition of otolith using laser ablation ICPMS. J. Fish. Aquat. Sci. Vol. 52, 1995.}
\item{Milton & Chenery. The effect of otolith storage methods on the concentrations of elements detected by laser-ablation ICPMS. J. of Fish Biology, Vol. 53, 1998.}
\item{Thorrold et al. 1998. Accurate classification of juvenile weakfish Cynoscion regalis to estuarine nursery areas based on chemical signatures in otoliths. Marine Ecology Press Series, Vol. 173, 1998.}
}
}
\examples{
#runElementR()
}
