\name{runElementR}
\alias{runElementR}
\title{
runElementR
}
\description{
Launch the shiny application contained in the elementR package.    
}
\usage{
runElementR()
}

\details{
By running \code{runElementR()}, user has access through its web browser opened as soon as the application is launched to the whole functionalities for reducing data from solid-phase ICPMS analysis (project creation or edition, data or project export, standard and sample filtration, verification of the machine drift). 
}


